requirements/base.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    make upgrade
#
amqp==5.3.1
    # via kombu
aniso8601==10.0.1
    # via tincan
apache-libcloud==3.8.0
    # via edx-event-routing-backends
appdirs==1.4.4
    # via fs
asgiref==3.10.0
    # via django
attrs==25.4.0
    # via openedx-events
billiard==4.2.2
    # via celery
celery==5.5.3
    # via
    #   edx-celeryutils
    #   event-tracking
    #   openedx-completion-aggregator
certifi==2025.10.5
    # via requests
cffi==2.0.0
    # via
    #   cryptography
    #   pynacl
charset-normalizer==3.4.4
    # via requests
click==8.3.0
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
    #   code-annotations
    #   edx-django-utils
click-didyoumean==0.3.1
    # via celery
click-plugins==1.1.1.2
    # via celery
click-repl==0.3.0
    # via celery
code-annotations==2.3.0
    # via edx-toggles
cryptography==46.0.3
    # via
    #   django-fernet-fields-v2
    #   pyjwt
django==4.2.25
    # via
    #   -c https://raw.githubusercontent.com/edx/edx-lint/master/edx_lint/files/common_constraints.txt
    #   -r requirements/base.in
    #   django-config-models
    #   django-crum
    #   django-fernet-fields-v2
    #   django-model-utils
    #   django-redis
    #   django-waffle
    #   djangorestframework
    #   drf-jwt
    #   edx-celeryutils
    #   edx-completion
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-event-routing-backends
    #   edx-toggles
    #   event-tracking
    #   jsonfield
    #   openedx-completion-aggregator
    #   openedx-events
    #   openedx-filters
django-config-models==2.9.0
    # via edx-event-routing-backends
django-crum==0.7.9
    # via
    #   edx-django-utils
    #   edx-toggles
django-fernet-fields-v2==0.9
    # via edx-event-routing-backends
django-model-utils==5.0.0
    # via
    #   -r requirements/base.in
    #   edx-celeryutils
    #   edx-completion
    #   openedx-completion-aggregator
django-object-actions==5.0.0
    # via -r requirements/base.in
django-redis==6.0.0
    # via edx-event-routing-backends
django-waffle==5.0.0
    # via
    #   edx-django-utils
    #   edx-drf-extensions
    #   edx-toggles
djangorestframework==3.16.1
    # via
    #   -r requirements/base.in
    #   django-config-models
    #   drf-jwt
    #   edx-completion
    #   edx-drf-extensions
    #   openedx-completion-aggregator
dnspython==2.8.0
    # via pymongo
drf-jwt==1.19.2
    # via edx-drf-extensions
edx-ccx-keys==2.0.2
    # via openedx-events
edx-celeryutils==1.4.0
    # via
    #   edx-event-routing-backends
    #   openedx-completion-aggregator
edx-completion==4.9
    # via openedx-completion-aggregator
edx-django-utils==8.0.1
    # via
    #   -r requirements/base.in
    #   django-config-models
    #   edx-drf-extensions
    #   edx-toggles
    #   event-tracking
    #   openedx-events
edx-drf-extensions==10.6.0
    # via edx-completion
edx-event-routing-backends==9.3.6
    # via openedx-completion-aggregator
edx-opaque-keys[django]==3.0.0
    # via
    #   -r requirements/base.in
    #   edx-ccx-keys
    #   edx-completion
    #   edx-drf-extensions
    #   openedx-completion-aggregator
    #   openedx-events
    #   openedx-filters
edx-toggles==5.4.1
    # via
    #   edx-completion
    #   edx-event-routing-backends
    #   event-tracking
    #   openedx-completion-aggregator
event-tracking==3.3.0
    # via
    #   edx-completion
    #   edx-event-routing-backends
    #   openedx-completion-aggregator
fastavro==1.12.1
    # via openedx-events
fasteners==0.20
    # via edx-event-routing-backends
fs==2.4.16
    # via xblock
idna==3.11
    # via requests
isodate==0.7.2
    # via edx-event-routing-backends
jinja2==3.1.6
    # via code-annotations
jsonfield==3.2.0
    # via
    #   edx-celeryutils
    #   edx-event-routing-backends
kombu==5.5.4
    # via celery
lxml==6.0.2
    # via xblock
mako==1.3.10
    # via xblock
markupsafe==3.0.3
    # via
    #   jinja2
    #   mako
    #   xblock
openedx-atlas==0.7.0
    # via -r requirements/base.in
openedx-completion-aggregator==4.2.0
    # via -r requirements/base.in
openedx-events==10.5.0
    # via event-tracking
openedx-filters==2.1.0
    # via edx-event-routing-backends
packaging==25.0
    # via kombu
pillow==12.0.0
    # via -r requirements/base.in
prompt-toolkit==3.0.52
    # via click-repl
psutil==7.1.1
    # via edx-django-utils
pycparser==2.23
    # via cffi
pyjwt[crypto]==2.10.1
    # via
    #   drf-jwt
    #   edx-drf-extensions
pymongo==4.15.3
    # via
    #   edx-opaque-keys
    #   event-tracking
pynacl==1.6.0
    # via edx-django-utils
python-dateutil==2.9.0.post0
    # via
    #   celery
    #   edx-event-routing-backends
    #   xblock
python-slugify==8.0.4
    # via code-annotations
pytz==2025.2
    # via
    #   edx-completion
    #   edx-event-routing-backends
    #   event-tracking
    #   tincan
    #   xblock
pyyaml==6.0.3
    # via
    #   code-annotations
    #   xblock
redis==6.4.0
    # via django-redis
requests==2.32.5
    # via
    #   apache-libcloud
    #   edx-drf-extensions
    #   edx-event-routing-backends
semantic-version==2.10.0
    # via edx-drf-extensions
simplejson==3.20.2
    # via xblock
six==1.17.0
    # via
    #   edx-ccx-keys
    #   event-tracking
    #   fs
    #   openedx-completion-aggregator
    #   python-dateutil
sqlparse==0.5.3
    # via django
stevedore==5.5.0
    # via
    #   code-annotations
    #   edx-django-utils
    #   edx-opaque-keys
text-unidecode==1.3
    # via python-slugify
tincan==1.0.0
    # via edx-event-routing-backends
typing-extensions==4.15.0
    # via edx-opaque-keys
tzdata==2025.2
    # via kombu
urllib3==2.5.0
    # via requests
vine==5.1.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.14
    # via prompt-toolkit
web-fragments==3.1.0
    # via xblock
webob==1.8.9
    # via xblock
xblock==5.2.0
    # via
    #   edx-completion
    #   openedx-completion-aggregator

# The following packages are considered to be unsafe in a requirements file:
# setuptools
learning-credentials==0.2.4
